Merged in add-an-extra-hint-to-what-a-clutername-is (pull request #11)
Add an extra hint to what a clustername is

A cluster and a shield are essentially the same thing. It's possible that a consumer
of the cloudformation template wouldn't be aware of this equivalency. But altering
the description of the field this should make it clearer to the consumer of the
template.